This is another push that includes some login functionality updates
This add I was able to add some functionality for logging in, signing up, and verifying the users in the system. I also added a logout functionality. I am now moving on to the messages and profile pages functionality so that hopefully I can get them to display the proper data.